Move slack release modals to specific directory
